requirements.txt
# Core dependencies
numpy>=1.24.0
matplotlib>=3.7.0
scikit-learn>=1.2.0
scipy>=1.10.0
tqdm>=4.65.0
pytest>=7.3.1

# New visualization stack
streamlit>=1.28.0
numpy>=1.21.0
pandas>=1.3.0
plotly>=5.3.0
matplotlib>=3.4.0
scikit-learn>=1.0.0lit>=1.25.0
plotly>=5.15.0

# Optional (kept for compatibility)
pygame>=2.5.0
PyOpenGL>=3.1.0
PyOpenGL-accelerate>=3.1.0
